+++ intelligentstore/intelligentstore/user.go
@@ -2,12 +2,12 @@ package intelligentstore
 
 // User represents a system user
 type User struct {
-	ID             int    `json:"id"`
+	ID             int64  `json:"id"`
 	DisplayName    string `json:"displayName"`
 	HashedPassword string `json:"-"`
 }
 
 // NewUser creates a new user
-func NewUser(id int, name, username string) *User {
+func NewUser(id int64, name, username string) *User {
 	return &User{id, name, username}
 }
