+++ intelligentstore/intelligentstore/user_test.go
@@ -9,6 +9,6 @@ import (
 func Test_NewUser(t *testing.T) {
 	u := NewUser(3, "test å user", "test@example.com")
 
-	assert.Equal(t, 3, u.ID)
+	assert.Equal(t, int64(3), u.ID)
 	assert.Equal(t, "test å user", u.DisplayName)
 }
